requirements.txt
torch
openai-whisper